Shuffling of TweetInfo in incorrect. fixes #23
Shuffling of the Tweets Against Users has been done in the incorrect
part of the game object. With this commit the backend part of the
project is done. Additional logging can be done later.

To get the game object from the backend a simple AJAX call must be made
to gameObject.php