GameObject doesn't repeat same Twitter User's Tweet, fixes #25
* There was a bug where the Game object had same twitter user’s tweet
twice. This was happening because when we needed to grab a new twitter
user after the initial 10 (because one of them was faulty) then the
information about the ID’s of already used Twitter Users does not
exist. Because of that the same Twitter User was getting selected
twice. (It could actually be selected more than 2 times too)

* Fix: The array of Twitter User IDs (`$alreadyUsedIDs`)which indicates
what all IDs have been used for the game has been made a global
variable now. SO now that information is not lost when we reach out to
grab a new user after the initial 10.