Extend CRON Job ability to accommodate more than 180 Twitter Users, Resolves #58
Logic behind extending the cron job script:

1) Run the script
2) Check the date in the FLAGS Table in the Database
3) If the Date in the DATABASE is today's date then abort the script.
There is no need to refresh twitter Users as Date is always changed to
today's date when all the twitter Users for today have been
been refreshed.
4) If the dates dont match then grab 150 or less than 150 Twitter Users
that have not been refreshed today.
5) If no Unrefreshed Twitter Users are retreived from the database then
change the LastRefreshFlag in the database to be
today's date so that no more refreshing for today takes place.