Leader Board Player Count Set At Backend, Resolves #73
This is security feature. Earlier the count of the leaderboard size was
coming from the front end. This is not good from a security point of
view. Someone could quest the backend and get all the players with
other data fro the database if possible. So therefore it was crucial to
set the count at the backend.